Update to Zero 0.6 - 2/4
Authorization is now defined in schema.ts, which makes a lot more
sense.

This change moves the auth part of zero.config.ts to schema.ts.

The entrypoint is now called `defineAuthorization` not
`defineConfig` and the structure of the returned type has changed
slightly.

Also the imports changed a bit.